# docs/swagger.yaml
basePath: /
definitions:
...
  views.Swagger_Social_Media_Post:
    properties:
      name:
        example: String
        type: string
      profile_image_url:
        example: String
        type: string
      social_media_url:
        example: String
        type: string
    type: object
  views.Swagger_Social_Media_Put:
    properties:
      name:
        example: String
        type: string
      social_media_url:
        example: String
        type: string
    type: object
...
host: localhost:os.Getenv("PORT")
info:
  contact:
    email: hansparson013@gmail.com
    name: Hans Parson
  description: Final Project Golang
  termsOfService: http://swagger.io/terms/
  title: My Grams APP
  version: v1.0
paths:
...
  /socialmedias:
    get:
      consumes:
      - application/json
      description: Get Social_Media
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Get SocialMedia
      tags:
      - SOCIAL_MEDIA
    post:
      consumes:
      - application/json
      description: Post SocialMedia
      parameters:
      - description: Post Social Media
        in: body
        name: Social_Media
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_Social_Media_Post'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Post SocialMedia
      tags:
      - SOCIAL_MEDIA
  /socialmedias/{socialMediaId}:
    delete:
      consumes:
      - application/json
      description: Get Social_MEdia
      parameters:
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Get Social_Media
      tags:
      - SOCIAL_MEDIA
    put:
      consumes:
      - application/json
      description: Update Social_Media
      parameters:
      - description: Update Social Media
        in: body
        name: Social_Media
        required: true
        schema:
          $ref: '#/definitions/views.Swagger_Social_Media_Put'
      - description: 'Token Barier example: ''Bearer 12355f32r'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Social Media ID
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Update Social_Media
      tags:
      - SOCIAL_MEDIA
...
swagger: "2.0"